+++ sql/llx_mahnwesen_interest_rate.key.sql
@@ -1,0 +1,1 @@
+ALTER TABLE llx_mahnwesen_interest_rate ADD UNIQUE INDEX uk_mahnwesen_interest_rate_from (entity, date_from);